Simple Alexa skill for text message alerts, DB bugfixes, added in Benry's account pages.

+++ SPECS_Web_Server/SPECS_Web_Server/Models/MedicalSensorData.cs
@@ -8,7 +8,7 @@ namespace SPECS_Web_Server.Models
     {
         [JsonIgnore]
 
-        public int userID { get; set; }
+        public long userID { get; set; }
 
         public float SpO2 { get; set; }
